+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.1
 
 - fixes bug that analyses results were not shown for shared surfaces (#186)
+- fixes bug that single topographies couldn't be selected in surface list (#190)
 - in order to find user names you have to type at least 3 characters (#184)
 - fixes highlighting and breadcrumb for "Sharing" page in navigation bar (#187)
 - improves display of help texts in topography forms (#180)